requirements2.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1603893133560/work
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828452693/work
asgiref==3.3.1
astor==0.8.1
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1600699180754/work
chardet==3.0.4
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cycler==0.10.0
Cython==3.0a6
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1602083700509/work
decorator==4.4.2
defusedxml==0.6.0
dill==0.2.7.1
Django==3.1.3
django-redis==4.12.1
djangorestframework==3.12.2
entrypoints==0.3
fire==0.3.1
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1597424467487/work
h5py @ file:///tmp/build/80754af9/h5py_1593454119955/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
image==1.5.33
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206598566/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1604101195213/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
jedi @ file:///tmp/build/80754af9/jedi_1596490743326/work
Jinja2==2.11.2
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014532738/work
Markdown @ file:///tmp/build/80754af9/markdown_1605111062842/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1594371495811/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603376012865/work
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.0
msgpack-numpy @ file:///tmp/build/80754af9/msgpack-numpy_1605111757211/work
mysqlclient==2.0.1
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914821128/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1605115881283/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
notebook @ file:///tmp/build/80754af9/notebook_1601501580008/work
numpy==1.19.4
olefile==0.46
opencv-python==4.4.0.46
packaging==20.4
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120451932/work
parse==1.18.0
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1594383317248/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1594384075987/work
pika==1.1.0
Pillow @ file:///tmp/build/80754af9/pillow_1603822253009/work
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.6.0
pycocotools==2.0.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
pyparsing==2.4.7
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141707582/work
python-dateutil==2.8.1
pytz==2020.4
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658308664/work
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
redis==3.5.3
requests==2.25.0
scikit-image==0.17.2
scipy @ file:///tmp/build/80754af9/scipy_1597686620742/work
Send2Trash==1.5.0
six @ file:///tmp/build/80754af9/six_1605205313296/work
slidingwindow==0.0.14
sqlparse==0.4.1
tabulate==0.8.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorpack==0.10.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-pose==0.1.1
tifffile==2020.10.1
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
tornado==6.0.4
tqdm==4.51.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
urllib3==1.26.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
youtube-dlc==2020.11.11.post3
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
